+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - maxSelected option
 - searchDebounce option
 - strings option
+- reset() method
 - eslint config
 - Message when not items selected
 - Selector loading spinner